refactor: 기존의 UserInput 클래스 내부에 입력값을 검증하던 메소드들은 모두 validator 패키지로 옮김
