+++ src/main/java/racingcar/validator/Validator.java
@@ -1,0 +1,7 @@
+package racingcar.validator;
+
+public abstract class Validator {
+
+    public abstract void validate(String input);
+
+}
